Add all functions to the app
Able to add books that are displayed as cards. The plus button will
bring up a form positioned in the middle of the screen and the rest
of the page are dimmed to make the form stand out. The read status is
displayed as a button and toggles read and not read.